+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 24/04/2020
+
+- Prevent loading data for areas that don't have a geostore.
+
 # 1.0.0
 
 ## 09/04/2020